concat hostName and port for ldap_connect
Instead of using two parameters for ldap_connect (hostName and port), this PR combines them to a single parameter. The docs of [ldap_connect](https://www.php.net/manual/en/function.ldap-connect.php) describe the host parameter as either a hostName or a LDAP URI. In case the latter is used, the second parameter (port) is ignored. When using a protocol, the parameters are combined, otherwise it stays the same.