feature: support nested inclusions
Previously we could only support one level of page inclusion. So, if we
have a three pages with this content:

    Page one: 'Test'

    Page two: 'Include page one: {{@1}}'

    Page three: 'Include page two: {{@2}}'

Rendering 'Page three' would not expand the nested includes, and would
result in:

    Include page two: Include page one: {{@1}}

To fix this we just keep expanding the includes until there are none
left.

Extract out the regex that matches the includes into a constant - since
this is now used in three places.