Add id tag to the registration form
Add an id tag to the registration form. An id tag already exists on the login form and this will allow both the login form and registration form elements be alterable from the "Custom HTML Head Content" option in Settings. We hide both of these forms because all of our users use federated authentication to log in. We do not want users creating accounts in the app via the registration form.